Pattern matching and collecting implicit records
Sections currently assume that all the names they use are in their
record, and nothing in inherited from the parent scope, but contrary to
that, all names are considered global in terms of the names of the
records and therefore what fields the records will have.

This needs some thought.